config: fix deploy targets
I moved staging.fosdem.org to apeiron (where it belongs).  We now use a
shared user (www-data) to own the website content.  To allow a user to
deploy content, just add their principal to /var/www/.k5login.

We currently use the same shared user for both staging and production.
If ever we want to restrict production to a subset of users, this can
be trivially accomplished by adding another account with a different
set of principals in its $HOME/.k5login.

Note that rsync -a will preserve permissions.  Users are recommended
to have a umask 0002.  Since we use a shared owner account though, it
doesn't really matter much.