schedule: first valiant attempt to improve readability
Try to limit (source code) lines to 80 characters so they fit in a
terminal window.  Adopt the LaTeX convention of inserting linebreaks
at the end of every sentence so that (future) content diffs become
more readable.  Fix some awkward English sentences here and there.

Apologies for the huge diff.  It's all for the best though. :-)