pentabarf: hack around security track and devroom
We have both a security main track and a security developer room this
year.  Of course this confuses things ...  Hack around the problem with
a hardcoded exception.  The proper / generic solution would be to track
the track names we've already seen and make an exception dynamically.
Or just not have a main track and a developer room with the same name!