pentabarf: nil image create/modify date
The image creation and modification date was set to the current
timestamp. Exports produced at different times would thus result in
different checksums for the image files, which in turn results in
different filenames being generated by nanoc when compiling the site.

Setting them to nil makes sure that data is not included in the image
at all, and that the image blobs are identical across exports.