Build the website in github actions
Idea is that more people can contribute to the website and build
without having to setup a whole pipeline. Once pretalx publishes its
yaml we can have a second test on that instead of the demo version.

The container was built and pushed to my package repository because
it seems github does not allow pushing packages when using
fine grained pats ( https://github.com/github/roadmap/issues/558 ).
If someone can get it at FOSDEM, that seems a better place.